BUG21873666 Fix max_connections and connect_timeout usage
We now correctly use the values passed in the configuration of Router
for max_connections and connect_timeout. Previously, setting the
max_connections or conenct_timeout options did not have any effect.

+++ src/routing/include/mysqlrouter/routing.h
@@ -33,7 +33,7 @@ namespace routing {
 extern const int kDefaultWaitTimeout;
 
 /** @brief Max number of active routes for this routing instance */
-extern const uint16_t kDefaultMaxConnections;
+extern const int kDefaultMaxConnections;
 
 /** @brief Timeout connecting to destination (in seconds)
  *
